requirements.txt
streamlit-webrtc
pydub
openpyxl
streamlit
openai-whisper
torch
gtts
pandas
numpy
soundfile
